Revert part of aa1cc7089a70d134f2d71f84293628e05818b07f
tokio-rs/tracing pulls in a lot more dependencies that log, also potentially
impacting negatively the build time of the crate.
While this was an acceptable choice for a single user of the crate, which was
what I had in mind when I made this commit in the first place, it is not
perhaps not a sensible choice for a project that aims at gaining tractiong and
seeks to provide a simple, safe and abstract wrapper for libnftnl.
Hence the move back to the log crate.